* commit '727e27340b94653c3994f6e40c9d4db8edaf1e6d':
  successfully linked api with controller
  RealAPI changes
  realapi update

+++ src/main/java/ca/uwo/csd/cs2212/team12/DataEntry.java
@@ -7,7 +7,7 @@ public class DataEntry implements Serializable {
 	 * The data entries.
 	 */
 	private int calBurned;
-	private int distanceTravelled;
+	private float distanceTravelled;
 	private int floorsClimbed;
 	private int stepsTaken;
 	private int actMins;
@@ -25,7 +25,7 @@ public class DataEntry implements Serializable {
 	 * @param sedMinutes This is the user's number of sedentary minutes.
 	 * @param theDate This is the date at which the previous values were recorded.
 	 */
-	public DataEntry(int cal, int dist, int floors, int steps, int actMinutes, int sedMinutes, String theDate){
+	public DataEntry(int cal, float dist, int floors, int steps, int actMinutes, int sedMinutes, String theDate){
 			 calBurned= cal;
 			 distanceTravelled= dist;
 			 floorsClimbed= floors;
@@ -60,14 +60,14 @@ public class DataEntry implements Serializable {
 	* This method returns the distance travelled by the user.
 	* @return int This is the value of distanceTravelled
 	*/
-	public int getDistanceTravelled() {
+	public float getDistanceTravelled() {
 		return distanceTravelled;
 	}
 
 	/**
 	* This method sets the distance travelled by the user.
 	*/
-	public void setDistanceTravelled(int distanceTravelled) {
+	public void setDistanceTravelled(float distanceTravelled) {
 		this.distanceTravelled = distanceTravelled;
 	}
